Removing unnecessary casts in interpreter instructions.

+++ src/System.Linq.Expressions/src/System/Linq/Expressions/Interpreter/DivInstruction.cs
@@ -68,7 +68,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (long)((long)l / (long)r);
+                    frame.Data[frame.StackIndex - 2] = (long)l / (long)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -106,7 +106,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (uint)((uint)l / (uint)r);
+                    frame.Data[frame.StackIndex - 2] = (uint)l / (uint)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -125,7 +125,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (ulong)((ulong)l / (ulong)r);
+                    frame.Data[frame.StackIndex - 2] = (ulong)l / (ulong)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -144,7 +144,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (float)((float)l / (float)r);
+                    frame.Data[frame.StackIndex - 2] = (float)l / (float)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -250,7 +250,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (long)((long)l % (long)r);
+                    frame.Data[frame.StackIndex - 2] = (long)l % (long)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -288,7 +288,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (uint)((uint)l % (uint)r);
+                    frame.Data[frame.StackIndex - 2] = (uint)l % (uint)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -307,7 +307,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (ulong)((ulong)l % (ulong)r);
+                    frame.Data[frame.StackIndex - 2] = (ulong)l % (ulong)r;
                 }
                 frame.StackIndex--;
                 return 1;
@@ -326,7 +326,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (float)((float)l % (float)r);
+                    frame.Data[frame.StackIndex - 2] = (float)l % (float)r;
                 }
                 frame.StackIndex--;
                 return 1;
